Split symbol table definitions
